+++ package/sdk/src/funcs/petsCreatePets.ts
@@ -5,6 +5,7 @@
 import { PetstoreCore } from "../core.js";
 import { encodeJSON } from "../lib/encodings.js";
 import * as M from "../lib/matchers.js";
+import { compactMap } from "../lib/primitives.js";
 import { safeParse } from "../lib/schemas.js";
 import { RequestOptions } from "../lib/sdks.js";
 import { pathToFunc } from "../lib/url.js";
@@ -52,10 +53,10 @@ export async function petsCreatePets(
 
   const path = pathToFunc("/pets")();
 
-  const headers = new Headers({
+  const headers = new Headers(compactMap({
     "Content-Type": "application/json",
     Accept: "application/json",
-  });
+  }));
 
   const context = {
     operationID: "createPets",
@@ -72,6 +73,7 @@ export async function petsCreatePets(
 
   const requestRes = client._createRequest(context, {
     method: "POST",
+    baseURL: options?.serverURL,
     path: path,
     headers: headers,
     body: body,
@@ -104,7 +106,8 @@ export async function petsCreatePets(
     | ConnectionError
   >(
     M.nil(201, components.ErrorT$inboundSchema.optional()),
-    M.fail(["4XX", "5XX"]),
+    M.fail("4XX"),
+    M.fail("5XX"),
     M.json("default", components.ErrorT$inboundSchema.optional()),
   )(response);
   if (!result.ok) {